Add explicit cast to fix Windows build.
Different defaulting rules on Windows means the test would fail otherwise.

PiperOrigin-RevId: 425682824

+++ tensorflow/python/ops/ragged/ragged_dispatch_test.py
@@ -908,7 +908,8 @@ class RaggedDispatchTest(test_util.TensorFlowTestCase, parameterized.TestCase):
       dict(
           op=array_ops.bitcast,
           kwargs={
-              'input': ragged_factory_ops.constant_value([[1, 2], [-1]]),
+              'input': ragged_factory_ops.constant_value([[1, 2], [-1]],
+                                                         dtype=dtypes.int64),
               'type': dtypes.uint64
           },
           expected=ragged_factory_ops.constant_value([[1, 2], [-1]],
